Entity iteration refactoring
Custom Iterator is used for Entity \JsonSerializable, \Serializable, \IteratorAggregate implementation and for toArray.

The main reason is to provide one consistent solution for them (toArray, getIterator and serializeToJson have the same logic). Easily configurable and replecable.

Public properties are back for optional backward compatibility.

Is possible to configure handlin null values on computed and public properties.